+++ src/modules/nft-manager/components/NftManagerDetails.vue
